Add files via upload
This code uses the pyzbar and PIL libraries to decode a QR code from the image github.png and print the data contained in the QR code.

Additionally, it is important to note that the code is using the decode() function from the pyzbar library to decode the QR code and the decode('ascii') function to convert the decoded data into ASCII format. Make sure to have the pyzbar and PIL libraries installed before running the code.